requirements.txt
langchain-openai==0.1.8
python-dotenv==1.0.1
langchain==0.2.1
svgwrite==1.4.3
langchain-nvidia-ai-endpoints==0.1.0
langchain-community==0.2.2

networkx==3.3
matplotlib==3.9.0

langgraph==0.0.64

langchainhub==0.1.17 